requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.0.7
django-filter==2.4.0
djangorestframework==3.11.0
idna==2.10
img==2.5
Pillow==8.1.0
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
